feat(AttemptCount) : 시도횟수 원자값 포장
