+++ src/test/java/racingcar/domain/AttemptCountTest.java
@@ -1,0 +1,2 @@
+package racingcar.domain;public class AttemptCountTest {
+}
